requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
e2b-code-interpreter>=1.0.1
tavily-python>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0